+++ CHANGELOG.md
@@ -7,3 +7,4 @@ This mod is distributed under Version 2 of the GPL.  A copy of the GPL is includ
 ## v3.0.7 (Oct, XX, 2022)
 
 - add mounts into raid AC tables
+- add Missing Venomhide Ravasaur and Sea Turtle in Wrath mounts #262